requirements.txt
# LangChain and related packages
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-experimental==0.0.25
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.6

# Google AI packages
google-generativeai==0.8.4
google-api-core==2.24.1
google-auth==2.38.0

# Database packages
pymysql==1.1.1
sqlalchemy==2.0.38
mysql-connector-python==9.2.0

# Embeddings and vector stores
sentence-transformers==3.4.1
chromadb==0.6.3
faiss-cpu==1.7.4

# Additional utilities
python-dotenv==1.0.0
pydantic==2.10.6
numpy==1.26.4
pandas==2.2.3
# protobuf==4.25.6
protobuf==3.20.3

# Web
cryptography==44.0.1
streamlit==1.22.0
tiktoken==0.4.0